vps_console_alloc() calls pts_alloc() which calls tty_makedev() and adds the newly allocaed *tp to the end of the tty_list tailq. vps_console_alloc() then manually calls tty_makedev() again with the proper cred and name to create the console.  This however happens on the same *tp. When tty_makedev() adds the *tp to the tty_list tailq a second time, the next pointer gets set to itself and the end of the list has a duplicate entry, with both elements pointing to themselves as next.  That violates the assertion that the last element must have a NULL next pointer.
Manually undo this for now with a helper function, removing the first entry
before re-adding it a second time.  That saves us from the panic.   This was
good as a proof-of-concept fix.   The proper solution to be implemented next
would be to only make pts_alloc() call tty_makedev() if requested (and
keep that as default), so that we can do it manually ourselves later.  That
also saves use one alloc/free cycle on an object we never need or intend to
have and use.