VPS restore has the requirement that snapshot must be page-aligned!
This is a userspace supplied malloc()ed buffer, which we cannot
guarantuee to be page-aligned.
We changed vpsctl to over-allocate and memmove the data if needed.
With that this works currently and makes sure the user space pages
inside the buffer are actually page-aligned.

It does not feel long-term stable but needs to be good enough for now.
Even with the check commented out restore did work to some extend
that we got a follow-up panic which more likely was related to state
not properly saved.