Add a lot more ERRMSG calls in order to better figure out where an error comes from.  Improve some error messages to easily identify them by adding line numbers.
Properly handle some error cases to unlock and/or cleanup
rather than directly returning.

Initialize more struct proc fields (some for now with a
"default" value and an XXX comment to deal with.

Disable CAPMODE while restoring (looking up) certain values
on the new system again, as otherwise we are not allowed to
(anymore).  The question is are we missing more saved state
instead?